+++ Mesh/meshGFaceBamg.cpp
@@ -119,9 +119,9 @@ void meshGFaceBamg(GFace *gf)
   std::vector<MVertex *> newVert;
   Triangle2 *bamgTriangles = new Triangle2[gf->triangles.size()];
   for(std::size_t i = 0; i < gf->triangles.size(); i++) {
-    int nodes[3] = {gf->triangles[i]->getVertex(0)->getIndex(),
-                    gf->triangles[i]->getVertex(1)->getIndex(),
-                    gf->triangles[i]->getVertex(2)->getIndex()};
+    int nodes[3] = {(int)gf->triangles[i]->getVertex(0)->getIndex(),
+                    (int)gf->triangles[i]->getVertex(1)->getIndex(),
+                    (int)gf->triangles[i]->getVertex(2)->getIndex()};
     double u1(bamgVertices[nodes[0]][0]);
     double u2(bamgVertices[nodes[1]][0]);
     double u3(bamgVertices[nodes[2]][0]);
@@ -149,8 +149,8 @@ void meshGFaceBamg(GFace *gf)
   for(std::vector<GEdge *>::const_iterator it = edges.begin();
       it != edges.end(); ++it) {
     for(std::size_t i = 0; i < (*it)->lines.size(); ++i) {
-      int nodes[2] = {(*it)->lines[i]->getVertex(0)->getIndex(),
-                      (*it)->lines[i]->getVertex(1)->getIndex()};
+      int nodes[2] = {(int)(*it)->lines[i]->getVertex(0)->getIndex(),
+                      (int)(*it)->lines[i]->getVertex(1)->getIndex()};
       bamgBoundary[count].init(bamgVertices, nodes, (*it)->tag());
       bamgBoundary[count].lab = count;
       count++;
